requirements.txt
playwright
pytest
pytest-bdd
pytest-xdist
pytest-rerunfailures
allure-pytest